fix(conversions): wide TryFrom<u128> rejects values past signed storage range (Int<2>); wide to_u64/to_u128 guard negative sign before truncating divide; migrate d_unified_foundation + num_traits tests

+++ tests/d_unified_foundation.rs
@@ -36,9 +36,12 @@ fn copy_and_clone() {
 
 #[test]
 fn equality_by_raw_storage() {
-    let a: D<i128, 5> = D(123);
-    let b: D<i128, 5> = D(123);
-    let c: D<i128, 5> = D(124);
+    // Uses `i64` storage: `Debug` (needed by assert_eq!/assert_ne!) is
+    // emitted per-storage by `decl_decimal_display!`, and `i128` is no
+    // longer a decimal storage type (D38 now backs onto `Int<2>`).
+    let a: D<i64, 5> = D(123);
+    let b: D<i64, 5> = D(123);
+    let c: D<i64, 5> = D(124);
     assert_eq!(a, b);
     assert_ne!(a, c);
 }
